#137 - Created a unit test for batman
