--- Source/grassroots/Batman.py
@@ -1,13 +1,0 @@
-class Batman:
-
-    def __init__(self):
-        pass
-
-    def serve_justice(self) -> bool:
-        bad_guy_in_jail = False
-        batman_exists = True
-
-        if batman_exists:
-            bad_guy_in_jail = True
-
-        return bad_guy_in_jail
