+++ Source/grassroots/tests/unitTests/test_batman.py
@@ -1,5 +1,19 @@
 from django.test import TestCase
-from Source.grassroots import Batman
+
+
+class Batman:
+
+    def __init__(self):
+        pass
+
+    def serve_justice(self) -> bool:
+        bad_guy_in_jail = False
+        batman_exists = True
+
+        if batman_exists:
+            bad_guy_in_jail = True
+
+        return bad_guy_in_jail
 
 
 class TestBatman(TestCase):
